Clean up API for plotting (`.mpl` -> `.plotting`)
