+++ tests/test_mesh_plots.py
@@ -6,7 +6,7 @@ import nanomesh
 from nanomesh import Mesher2D
 from nanomesh.mesh import TriangleMesh
 from nanomesh.mesh2d import compare_mesh_with_image
-from nanomesh.mpl.meshplot import plot_line_triangle
+from nanomesh.plotting.meshplot import line_triangle_plot
 
 
 @image_comparison2(baseline_images=['line_mesh'])
@@ -17,7 +17,7 @@ def test_line_mesh_plot(line_tri_mesh):
 
 @image_comparison2(baseline_images=['line_triangle_plot'])
 def test_line_triangle_mesh_plot(line_tri_mesh):
-    plot_line_triangle(line_tri_mesh)
+    line_triangle_plot(line_tri_mesh)
 
 
 @image_comparison2(baseline_images=['triangle_mesh'])
